fix(cli): bound MCP attach replacement teardown
Pass slot cancellation into RMCP initialization and reject reconnect admission when predecessor teardown exceeds its bound.
