+++ changes/1677.fixed.md
@@ -7,4 +7,6 @@ cap the gateway LRU-evicts an idle slot and fail-fasts if every slot is
 still active. `mcp gc` reaps orphaned `astrid`/`aos mcp attach`
 processes whose parent is dead, and never signals Python
 `aos-mcp-frame`; unrelated wrapper commands are ignored. `mcp serve`
-remains the explicit per-session escape hatch. Closes #1677
+remains the explicit per-session escape hatch. Replacement teardown now
+cancels RMCP initialization and rejects a reconnect when the old slot does
+not finish within its bounded teardown window. Closes #1677
